LayoutTests/http/tests/preload/not_delaying_window_onload_before_discovery-expected.txt

PASS finishedScript is false
PASS finishedImage is false
PASS successfullyParsed is true

TEST COMPLETE
PASS windowOnLoad is true
PASS windowOnLoad is true
